implement the Eslint and prettier
