+++ ambillion-ui/.vscode/settings.json
@@ -1,0 +1,13 @@
+{
+    "editor.formatOnSave": true,
+    "editor.defaultFormatter": "esbenp.prettier-vscode",
+    "[javascript]": {
+      "editor.formatOnSave": true
+    },
+    "eslint.validate": [
+      "javascript",
+      "javascriptreact",
+      { "language": "javascript", "autoFix": true },
+      { "language": "javascriptreact", "autoFix": true }
+    ]
+  }
